requirements.txt
bs4==0.0.1
datasets==2.13.1
huggingface-hub==0.16.4
pandas==1.3.5
requests-oauthlib==1.3.1
tokenizers==0.13.3
torch==1.13.1
tqdm==4.65.0
transformers==4.30.2
